Refactor NetworkCoverageMap to use lazy loading for Toaster and dynamic imports for html2canvas and toast; add exportUtils for PDF and CSV generation with comprehensive data formatting and structure
